LayoutTests/fast/regions/full-screen-video-from-region-expected.txt

This tests the full screen video playing from a region flow. After entering full screen mode, only the full screen video should be visible. Click go full screen to run the test.
EVENT(webkitfullscreenchange)
END OF TEST
